mesh/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mesh"
version = "0.1.0"
description = "Decentralized Multi-Agent Supply Chain Coordination on Tashi Vertex"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "PyNaCl>=1.5.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["../tests"]
markers = [
    "unit: fast isolated tests",
    "integration: tests requiring MQTT broker",
    "chaos: chaos engineering tests",
]
